as discussed in oasp/oasp4j#219, I refactored the offermanagement again to the use case pattern
